fix(pr-workflow): close final critic findings on cap ratchet and summary growth
Final critic review of 6ce3801 returned NEEDS_REVISION. Two substantive
findings, both correct.

C1 (blocker) - 6ce3801 introduced a CI-breaking FR-006 ratchet violation.
tests/unit/hooks/tool-summarizer-exempt.adversarial.test.ts grew from 1194 to
1203 lines. That file is far over the 500-line cap, so the ratchet forbids any
growth, and .github/workflows/ci.yml runs the check with enforcement on. The
branch could not have passed CI. Resolved by extracting the floor-contract
block into tool-summarizer-exempt-floor.test.ts rather than trimming comments,
so the over-cap file drops to 1100 lines - below where it started - instead of
merely holding steady. No assertion was weakened; all three tests moved
verbatim.

The same finding also caught a false completion claim: 08-test-results.md
asserted the ratchet was clean, measured against b4534a5 and never re-run
after 6ce3801. The artifact now records the correction and the lesson - a gate
result is evidence only for the tree it was run against.

C2 - stage A persisted full check output to .swarm/summaries/ on every run,
success and failure. cleanupSummaries has zero callers and retention_days has
no reader, so nothing reclaims that directory; a long feedback session with
large check batches grew it monotonically. Persistence now runs only on the
failure path, where the output has recovery value. A successful run returns
per-check summaries and persists nothing. The fail-open behavior on a
persistence throw is unchanged and still covered by its regression test.

Also corrected:
- A comment claiming the last results entry is always the check that triggered
  the failure. True only on the exit-code path; on the pre-check state-change
  and execution-throw paths the failing check has not been pushed yet.
- The release note omitted two user-observable changes: bounded
  argument-validation error lists, and the non-negative clamp on freed-token
  accounting.
- The cross-session de-dupe key collision was documented only in a
  non-shipping trace artifact; the caveat now lives on the deliveredLaneOutputs
  doc comment in source.
- swarm-pr-feedback/SKILL.md described Stage A persistence as unconditional,
  which is no longer true after C2.
