+++ tests/unit/hooks/tool-summarizer-exempt.adversarial.test.ts
@@ -1093,111 +1093,8 @@ describe('createToolSummarizerHook - Adversarial Tests for exempt_tools', () => 
 		});
 	});
 
-	/**
-	 * Attack Vector 11: retrieve_lane_output exemption ratchet
-	 * Verify retrieve_lane_output is treated like retrieve_summary — exempt by default,
-	 * not bypassable via suffix/prefix/case variations.
-	 */
-	describe('retrieve_lane_output exemption ratchet', () => {
-		it('should exempt retrieve_lane_output when in exempt list', async () => {
-			const config: SummaryConfig = {
-				enabled: true,
-				threshold_bytes: 1024,
-				max_summary_chars: 500,
-				max_stored_bytes: 1024 * 1024,
-				retention_days: 7,
-				exempt_tools: [
-					'retrieve_summary',
-					'retrieve_lane_output',
-					'task',
-					'read',
-				],
-			};
-
-			const hook = createToolSummarizerHook(config, tempDir);
-
-			const output = {
-				title: 'Test Tool',
-				output: largeOutput,
-				metadata: {},
-			};
-
-			await hook(
-				{ tool: 'retrieve_lane_output', sessionID: 'test', callID: 'test' },
-				output,
-			);
-
-			expect(output.output).toBe(largeOutput);
-		});
-
-		it('should NOT exempt retrieve_lane_output with suffix', async () => {
-			const config: SummaryConfig = {
-				enabled: true,
-				threshold_bytes: 1024,
-				max_summary_chars: 500,
-				max_stored_bytes: 1024 * 1024,
-				retention_days: 7,
-				exempt_tools: [
-					'retrieve_summary',
-					'retrieve_lane_output',
-					'task',
-					'read',
-				],
-			};
-
-			const hook = createToolSummarizerHook(config, tempDir);
-
-			const output = {
-				title: 'Test Tool',
-				output: largeOutput,
-				metadata: {},
-			};
-
-			await hook(
-				{
-					tool: 'retrieve_lane_output_extra',
-					sessionID: 'test',
-					callID: 'test',
-				},
-				output,
-			);
-
-			expect(output.output).not.toBe(largeOutput);
-		});
-
-		// Contract change: SUMMARIZER_EXEMPT_TOOL_NAMES (src/config/constants.ts)
-		// is now an unconditional FLOOR, not a default operator config can
-		// replace. `retrieve_lane_output` is a floor member — summarizing a lane
-		// payload would destroy the `output_ref` that is its only recovery path,
-		// so it must stay exempt even when the operator's exempt_tools list
-		// omits it entirely. This test previously asserted the OLD replacement
-		// semantics (missing from the list => summarized); it now asserts the
-		// floor holds. See docs/releases/pending/pr-workflow-lane-output-context.md.
-		it('should still exempt retrieve_lane_output (floor) even when missing from exempt list', async () => {
-			const config: SummaryConfig = {
-				enabled: true,
-				threshold_bytes: 1024,
-				max_summary_chars: 500,
-				max_stored_bytes: 1024 * 1024,
-				retention_days: 7,
-				exempt_tools: ['retrieve_summary', 'task', 'read'],
-			};
-
-			const hook = createToolSummarizerHook(config, tempDir);
-
-			const output = {
-				title: 'Test Tool',
-				output: largeOutput,
-				metadata: {},
-			};
-
-			await hook(
-				{ tool: 'retrieve_lane_output', sessionID: 'test', callID: 'test' },
-				output,
-			);
-
-			// Floor exemption applies regardless of operator exempt_tools content.
-			expect(output.output).toBe(largeOutput);
-		});
-	});
+	// Attack Vector 11 (retrieve_lane_output exemption ratchet, including the
+	// unconditional-floor contract test) moved to
+	// tests/unit/hooks/tool-summarizer-exempt-floor.test.ts — this file is
+	// already over the FR-006 500-line cap and must not grow further.
 });
